backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
PyJWT>=2.8.0
argon2-cffi>=23.1.0
passlib>=1.7.4
python-multipart>=0.0.9
Pillow>=10.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
anthropic>=0.40.0
jellyfish>=1.0.0
